move ClangHelpers
